added validation to transaction

+++ Api/ShoppingWebApi/src/core/db/migrations/1566656253270-transactions.ts
@@ -31,7 +31,7 @@ export class transactions1566656253270 implements MigrationInterface {
                     new TableColumn({name: 'id',generationStrategy:'increment', type: 'int',isPrimary: true,isGenerated: true,isUnique: true,isNullable: false}),
                     new TableColumn({name: 'date',type: 'datetime',isPrimary: false,isGenerated: false,isUnique: false,isNullable: false, default: 'CURRENT_TIMESTAMP'}),
                     new TableColumn({name: 'price',type: 'decimal',isPrimary: false,isGenerated: false,isUnique: false,isNullable: false,default:0.00}),
-                    new TableColumn({name: 'quantity', type: 'int',isPrimary: false,isGenerated: false,isUnique: false,isNullable: false}),
+                    new TableColumn({name: 'quantity', type: 'decimal',isPrimary: false,isGenerated: false,isUnique: false,isNullable: false}),
                     new TableColumn({name: 'productId', type: 'int',isPrimary: false,isGenerated: false,isUnique: false,isNullable: false}),
                     new TableColumn({name: 'transactionId', type: 'int',isPrimary: false,isGenerated: false,isUnique: false,isNullable: false}),
                 ]
